Minor unit test tweaks
Mostly just print statements to help me figure out which test is
hanging.

Did add a couple more arities to the get-byte-array-option function
wrapper for sockets, which helps with the principle of least surprise.

The tests still don't pass as regularly as they should, but it seems
like it might be more-or-less usable.