Unencrypted dealer (server) test mostly works
It throws an exception (missing resource) when I try to unbind that
socket, but at least the messages are exchanged.

Add a first draft of a more realistic version, where the router does the
binding. This version of that does not seem to be exchanging any
messages.

Really just checking this version in to move to a faster machine because
the startup time is highly annoying.