Ressurect the test that's trying to use ZAP
I'm receiving the AUTH request, and I think I'm sending back an
"authenticated" response (based on digging through the zeromq and czmq
source code). But I'm getting the request twice (which is an improvement
over just receiving it in an infinite loop), which seems worrisome.

And the actual messages aren't going through.