Debug my ZAP unit test a wee more
Changing the endpoint for the handler socket (so that it isn't used) let
the rest of the test run through.

It crashed the JVM during the first run through. I think that was
probably because I was trying to unbind the real ZAP end-point.

The second time through it ran fine.

Restoring it took me back to having it called over and over, which I
thought I'd managed to fix last night.

So 1 step forward, 2 back.