Add another encrypted push/pull test
Modelled after the faux-encrypted inproc version (except setting up the
client properly with the server's public key).

I'm not sure why it runs through correctly when the alternative version
is failing.

Move the test that does nothing but low-level CURVE setup/teardown into
the bindind-test ns, because it felt so lonely.