WIP - Debugging broken Socket spec
When I try to call s/valid? on one, I get an error about generation.

Actually, this version is just broken. I got interrupted in the middle
of trying to figure out how to declare a generator. Really just checking
it in because I'm sure I'll lose track of where I am before I get a
chance to get back to this.