Throw exception at socket spec generator
I'm getting there from trying to run (s/explain-data ::socket-spec
foo). I'm only calling that after I've checked that foo isn't valid.

This is mostly a sanity check to try to nail down the actual error
source. The stack trace is getting clipped inside the bowels of lazy
sequence handling before I can find anything interesting.

This is almost a desperation measure and shouldn't be committed. This
commit really shouldn't make it into production.

But I think it points to a possible bug in spec, so I want to make sure
I don't lose track of it yet again. The only time I seem to run across
this is when it's late at night and I'm desperate to get some sleep.