Have a working generator spec for read sockets
And probably a matching one for writers, though that one's really
boring.

Have a manual test to verify the reader, at any rate.

This seems pretty silly. But the entire point is to have a spec I can
use in frereth.common's event loops.

Though, honestly, this seems like a very useful thing to have for
exercising 0mq in general.